% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the location of the file}
}
\value{
a data frame as read by read_csv with default options.
}
\description{
A function for reading in FARS data, it checks if a given file exists. 
If it does then it reads it in using the read_csv function.
If it does not then it returns 'does not exist'
}
\examples{
fars_read("accident_2013.csv.bz2")

}
