% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of character vectors}
}
\value{
returns a list containing data (or an error) for each of the years passed in.
}
\description{
A wrapper function for applying the `fars_read` & `make_filename` to read multiple years of data using lapply.
If an invalid year string is provided it will return an error
}
\examples{
years <- c("2013","2014")
fars_read_years(years)

}
