% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a list of years (strings) to summarise data for.}
}
\value{
a tibble of the number of datapoints with columns for each year and rows for each month
}
\description{
A summary function to display the number of rows of data for each month of each year.
}
\examples{
years <- c("2013","2014")
fars_summarize_years(years)
}
