% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fars_data.R
\name{get_fars_data}
\alias{get_fars_data}
\title{get_fars_data}
\usage{
get_fars_data(force = FALSE,
  durl = "https://d3c33hcgiwev3.cloudfront.net/_e1adac2a5f05192dc8780f3944feec13_fars_data.zip?Expires=1546300800&Signature=byljyRPskWKpmmyUfzF1hMKln-IdQOoa4vCnVh82WqFBP4HgnTT~3kjr~XysTzFdVuS-6bTvLZ6V3b1jp3Cg4Gq797jTAGqK3NPfi1oRyDGm9MjBerNNIS8MiMcSJqFgInBXDxFXr-SOqFLE9GLUDNKoCPmb3qynYfZRV06FtQU_&Key-Pair-Id=APKAJLTNE6QMUY6HBC5A")
}
\arguments{
\item{force}{}
}
\description{
This function downloads the fars data to the local apps directory of the current user (if given permission to do so)
Function created to allow vignette to be built
}
\examples{
get_fars_data(force = TRUE)
}
