% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a character string defining a year}
}
\value{
returns a string in the format accident_YYYY.csv.bz2
}
\description{
This function takes a year as a character string, coerces it to an integer, then uses sprintf to generate a filename for one of the FARS datasets.
This function has no input sanitation and will allow any year string and any character.
}
\examples{
make_filename("2013")
}
