% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fars_data.R
\name{get_fars_data}
\alias{get_fars_data}
\title{get_fars_data}
\usage{
get_fars_data(force = FALSE,
  durl = "https://d3c33hcgiwev3.cloudfront.net/_e1adac2a5f05192dc8780f3944feec13_fars_data.zip?Expires=1546560000&Signature=gabwk0MIC8yBFZFr75x5P-SbDIu7suFq3gOm1Aef1U4SwQtPUavEm8o3YB107kJcPRZXf2ZPADrcISbSJ4TXnEN725ewt8YPZBajP0LtZCtBTqNA~P5UYbVd620rw4ew~ZnXqoBO1Ofr6hY0LP4Kv5ZEV2TtN75hgRX75JAyCcE_&Key-Pair-Id=APKAJLTNE6QMUY6HBC5A")
}
\arguments{
\item{force}{}
}
\description{
This function downloads the fars data to the local apps directory of the current user (if given permission to do so)
Function created to allow vignette to be built
}
\examples{
get_fars_data(force = TRUE)
}
